third_party/blink/web_tests/fast/canvas/bug1017918.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<canvas id="canvas1" width=974px></canvas>
<script>
test(() => {
  let C = document.getElementById("canvas1");
  let Z = C.getContext("2d");
  assert_throws_js(RangeError, () => Z.getImageData(174.696924716,
    -3.94829667221, 6.80556775521, 291126556.736));
}, "Regression test for https://crbug.com/1017918.");
</script>
